added api and create camper functionality
