+++ App/package.json
@@ -1,30 +1,21 @@
 {
   "name": "app",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "@reach/router": "^1.1.1",
-    "final-form": "^4.8.3",
-    "neo4j-driver": "^1.6.2",
-    "react": "^16.4.1",
-    "react-dom": "^16.4.1",
-    "react-final-form": "^3.6.4",
-    "react-scripts": "1.1.4",
-    "react-select": "^1.2.1",
-    "react-window": "^1.0.0-alpha.4"
-  },
+  "version": "1.0.0",
+  "description": "",
+  "main": "server.js",
   "scripts": {
-    "start": "react-scripts start",
-    "deploy": "npm run build && cd build && now --public --name camper-test-tracker -e NODE_ENV=\"production\"",
-    "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject",
-    "precommit": "pretty-quick --staged"
+    "start": "concurrently \"npm run server\" \"npm run client\"",
+    "server": "nodemon server.js",
+    "client": "cd client && npm run start"
   },
+  "author": "",
+  "license": "ISC",
   "devDependencies": {
-    "eslint-plugin-prettier": "^2.6.1",
-    "husky": "^0.14.3",
-    "prettier": "1.13.7",
-    "pretty-quick": "^1.6.0"
+    "concurrently": "^3.6.0",
+    "nodemon": "^1.18.2"
+  },
+  "dependencies": {
+    "fastify": "^1.8.0",
+    "neo4j-driver": "^1.6.2"
   }
 }
