+++ Queries/create_summers.cypher
@@ -1,0 +1,4 @@
+// Create Summers
+UNWIND {summers} as summer
+MERGE (s:Summer {year: summer.value})
+RETURN s
