+++ app/client/.eslintrc
