+++ app/client/public/favicon.ico
