+++ app/client/public/manifest.json
