+++ app/client/public/viking-ship.png
