+++ app/client/src/registerServiceWorker.js
